TOOLS/PERF: Split perf test into separated modules
