+++ src/tools/perf/Makefile.am
@@ -11,9 +11,17 @@
 SUBDIRS = cuda rocm lib
 CC      = $(UCX_PERFTEST_CC)
 
-noinst_HEADERS        = api/libperf.h
-bin_PROGRAMS          = ucx_perftest
-ucx_perftest_SOURCES  = perftest.c
+noinst_HEADERS = \
+	perftest.h \
+	api/libperf.h
+
+bin_PROGRAMS = ucx_perftest
+
+ucx_perftest_SOURCES = \
+	perftest.c \
+	perftest_run.c \
+	perftest_params.c
+
 ucx_perftest_CPPFLAGS = $(BASE_CPPFLAGS) $(RTE_CPPFLAGS)
 ucx_perftest_CFLAGS   = $(BASE_CFLAGS) $(OPENMP_CFLAGS)
 ucx_perftest_LDFLAGS  = $(RTE_LDFLAGS) -Wl,-dynamic-list-data
