+++ src/tools/perf/perftest.c
@@ -8,90 +8,7 @@
 * See file LICENSE for terms.
 */
 
-
-#ifdef HAVE_CONFIG_H
-#  include "config.h"
-#endif
-
-#include "api/libperf.h"
-#include "lib/libperf_int.h"
-
-#include <ucs/sys/string.h>
-#include <ucs/sys/sys.h>
-#include <ucs/sys/sock.h>
-#include <ucs/debug/log.h>
-
-#include <sys/socket.h>
-#include <arpa/inet.h>
-#include <stdlib.h>
-#include <stdio.h>
-#include <unistd.h>
-#include <netdb.h>
-#include <getopt.h>
-#include <string.h>
-#include <sys/types.h>
-#include <sys/poll.h>
-#include <locale.h>
-#if defined (HAVE_MPI)
-#  include <mpi.h>
-#elif defined (HAVE_RTE)
-#   include<rte.h>
-#endif
-
-#define MAX_BATCH_FILES         32
-#define MAX_CPUS                1024
-#define TL_RESOURCE_NAME_NONE   "<none>"
-#define TEST_PARAMS_ARGS        "t:n:s:W:O:w:D:i:H:oSCIqM:r:E:T:d:x:A:BUem:R:"
-#define TEST_ID_UNDEFINED       -1
-
-enum {
-    TEST_FLAG_PRINT_RESULTS = UCS_BIT(0),
-    TEST_FLAG_PRINT_TEST    = UCS_BIT(1),
-    TEST_FLAG_SET_AFFINITY  = UCS_BIT(8),
-    TEST_FLAG_NUMERIC_FMT   = UCS_BIT(9),
-    TEST_FLAG_PRINT_FINAL   = UCS_BIT(10),
-    TEST_FLAG_PRINT_CSV     = UCS_BIT(11)
-};
-
-typedef struct sock_rte_group {
-    int                          is_server;
-    int                          connfd;
-} sock_rte_group_t;
-
-
-typedef struct test_type {
-    const char                   *name;
-    ucx_perf_api_t               api;
-    ucx_perf_cmd_t               command;
-    ucx_perf_test_type_t         test_type;
-    const char                   *desc;
-    const char                   *overhead_lat;
-    unsigned                     window_size;
-} test_type_t;
-
-
-typedef struct perftest_params {
-    ucx_perf_params_t            super;
-    int                          test_id;
-} perftest_params_t;
-
-
-struct perftest_context {
-    perftest_params_t            params;
-    const char                   *server_addr;
-    int                          port;
-    int                          mpi;
-    unsigned                     num_cpus;
-    unsigned                     cpus[MAX_CPUS];
-    unsigned                     flags;
-
-    unsigned                     num_batch_files;
-    char                         *batch_files[MAX_BATCH_FILES];
-    char                         *test_names[MAX_BATCH_FILES];
-
-    sock_rte_group_t             sock_rte_group;
-};
-
+#include "perftest.h"
 
 test_type_t tests[] = {
     {"am_lat", UCX_PERF_API_UCT, UCX_PERF_CMD_AM, UCX_PERF_TEST_TYPE_PINGPONG,
@@ -223,385 +140,7 @@ static int safe_recv(int sock, void *data, size_t size,
     return sock_io(sock, recv, POLLIN, data, size, progress, arg, "recv");
 }
 
-static void print_progress(char **test_names, unsigned num_names,
-                           const ucx_perf_result_t *result, unsigned flags,
-                           int final, int is_server, int is_multi_thread)
-{
-    static const char *fmt_csv;
-    static const char *fmt_numeric;
-    static const char *fmt_plain;
-    unsigned i;
-
-    if (!(flags & TEST_FLAG_PRINT_RESULTS) ||
-        (!final && (flags & TEST_FLAG_PRINT_FINAL)))
-    {
-        return;
-    }
-
-    if (flags & TEST_FLAG_PRINT_CSV) {
-        for (i = 0; i < num_names; ++i) {
-            printf("%s,", test_names[i]);
-        }
-    }
-
-#if _OPENMP
-    if (!final) {
-        printf("[thread %d]", omp_get_thread_num());
-    } else if (flags & TEST_FLAG_PRINT_RESULTS) {
-        printf("Final:    ");
-    }
-#endif
-
-    if (is_multi_thread && final) {
-        fmt_csv     = "%4.0f,%.3f,%.2f,%.0f\n";
-        fmt_numeric = "%'18.0f %29.3f %22.2f %'24.0f\n";
-        fmt_plain   = "%18.0f %29.3f %22.2f %23.0f\n";
-
-        printf((flags & TEST_FLAG_PRINT_CSV)   ? fmt_csv :
-               (flags & TEST_FLAG_NUMERIC_FMT) ? fmt_numeric :
-                                                 fmt_plain,
-               (double)result->iters,
-               result->latency.total_average * 1000000.0,
-               result->bandwidth.total_average / (1024.0 * 1024.0),
-               result->msgrate.total_average);
-    } else {
-        fmt_csv     = "%4.0f,%.3f,%.3f,%.3f,%.2f,%.2f,%.0f,%.0f\n";
-        fmt_numeric = "%'18.0f %10.3f %9.3f %9.3f %11.2f %10.2f %'11.0f %'11.0f\n";
-        fmt_plain   = "%18.0f %10.3f %9.3f %9.3f %11.2f %10.2f %11.0f %11.0f\n";
-
-        printf((flags & TEST_FLAG_PRINT_CSV)   ? fmt_csv :
-               (flags & TEST_FLAG_NUMERIC_FMT) ? fmt_numeric :
-                                                 fmt_plain,
-               (double)result->iters,
-               result->latency.percentile * 1000000.0,
-               result->latency.moment_average * 1000000.0,
-               result->latency.total_average * 1000000.0,
-               result->bandwidth.moment_average / (1024.0 * 1024.0),
-               result->bandwidth.total_average / (1024.0 * 1024.0),
-               result->msgrate.moment_average,
-               result->msgrate.total_average);
-    }
-
-    fflush(stdout);
-}
-
-static void print_header(struct perftest_context *ctx)
-{
-    const char *overhead_lat_str;
-    const char *test_data_str;
-    const char *test_api_str;
-    test_type_t *test;
-    unsigned i;
-
-    test = (ctx->params.test_id == TEST_ID_UNDEFINED) ? NULL :
-           &tests[ctx->params.test_id];
-
-    if ((ctx->flags & TEST_FLAG_PRINT_TEST) && (test != NULL)) {
-        if (test->api == UCX_PERF_API_UCT) {
-            test_api_str = "transport layer";
-            switch (ctx->params.super.uct.data_layout) {
-            case UCT_PERF_DATA_LAYOUT_SHORT:
-                test_data_str = "short";
-                break;
-            case UCT_PERF_DATA_LAYOUT_SHORT_IOV:
-                test_data_str = "short iov";
-                break;
-            case UCT_PERF_DATA_LAYOUT_BCOPY:
-                test_data_str = "bcopy";
-                break;
-            case UCT_PERF_DATA_LAYOUT_ZCOPY:
-                test_data_str = "zcopy";
-                break;
-            default:
-                test_data_str = "(undefined)";
-                break;
-            }
-        } else if (test->api == UCX_PERF_API_UCP) {
-            test_api_str  = "protocol layer";
-            test_data_str = "(automatic)"; /* TODO contig/stride/stream */
-        } else {
-            return;
-        }
-
-        printf("+------------------------------------------------------------------------------------------+\n");
-        printf("| API:          %-60s               |\n", test_api_str);
-        printf("| Test:         %-60s               |\n", test->desc);
-        printf("| Data layout:  %-60s               |\n", test_data_str);
-        printf("| Send memory:  %-60s               |\n", ucs_memory_type_names[ctx->params.super.send_mem_type]);
-        printf("| Recv memory:  %-60s               |\n", ucs_memory_type_names[ctx->params.super.recv_mem_type]);
-        printf("| Message size: %-60zu               |\n", ucx_perf_get_message_size(&ctx->params.super));
-        if ((test->api == UCX_PERF_API_UCP) &&
-            (test->command == UCX_PERF_CMD_AM)) {
-            printf("| AM header size: %-60zu             |\n",
-                   ctx->params.super.ucp.am_hdr_size);
-        }
-    }
-
-    if (ctx->flags & TEST_FLAG_PRINT_CSV) {
-        if (ctx->flags & TEST_FLAG_PRINT_RESULTS) {
-            for (i = 0; i < ctx->num_batch_files; ++i) {
-                printf("%s,", ucs_basename(ctx->batch_files[i]));
-            }
-            printf("iterations,%.1f_percentile_lat,avg_lat,overall_lat,avg_bw,overall_bw,avg_mr,overall_mr\n", ctx->params.super.percentile_rank);
-        }
-    } else {
-        if (ctx->flags & TEST_FLAG_PRINT_RESULTS) {
-            overhead_lat_str = (test == NULL) ? "overhead" : test->overhead_lat;
-
-            printf("+--------------+--------------+------------------------------+---------------------+-----------------------+\n");
-            printf("|              |              |       %8s (usec)        |   bandwidth (MB/s)  |  message rate (msg/s) |\n", overhead_lat_str);
-            printf("+--------------+--------------+----------+---------+---------+----------+----------+-----------+-----------+\n");
-            printf("|    Stage     | # iterations | %4.1f%%ile | average | overall |  average |  overall |  average  |  overall  |\n", ctx->params.super.percentile_rank);
-            printf("+--------------+--------------+----------+---------+---------+----------+----------+-----------+-----------+\n");
-        } else if (ctx->flags & TEST_FLAG_PRINT_TEST) {
-            printf("+------------------------------------------------------------------------------------------+\n");
-        }
-    }
-}
-
-static void print_test_name(struct perftest_context *ctx)
-{
-    char buf[200];
-    unsigned i, pos;
-
-    if (!(ctx->flags & TEST_FLAG_PRINT_CSV) && (ctx->num_batch_files > 0)) {
-        strcpy(buf, "+--------------+--------------+----------+---------+---------+----------+----------+-----------+-----------+");
-
-        pos = 1;
-        for (i = 0; i < ctx->num_batch_files; ++i) {
-           if (i != 0) {
-               buf[pos++] = '/';
-           }
-           memcpy(&buf[pos], ctx->test_names[i],
-                  ucs_min(strlen(ctx->test_names[i]), sizeof(buf) - pos - 1));
-           pos += strlen(ctx->test_names[i]);
-        }
-
-        if (ctx->flags & TEST_FLAG_PRINT_RESULTS) {
-            printf("%s\n", buf);
-        }
-    }
-}
-
-static void print_memory_type_usage(void)
-{
-    ucs_memory_type_t it;
-    for (it = UCS_MEMORY_TYPE_HOST; it < UCS_MEMORY_TYPE_LAST; it++) {
-        if (ucx_perf_mem_type_allocators[it] != NULL) {
-            printf("                        %s - %s\n",
-                   ucs_memory_type_names[it],
-                   ucs_memory_type_descs[it]);
-        }
-    }
-}
-
-static void usage(const struct perftest_context *ctx, const char *program)
-{
-    static const char* api_names[] = {
-        [UCX_PERF_API_UCT] = "UCT",
-        [UCX_PERF_API_UCP] = "UCP"
-    };
-    test_type_t *test;
-    int UCS_V_UNUSED rank;
-
-#ifdef HAVE_MPI
-    MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-    if (ctx->mpi && (rank != 0)) {
-        return;
-    }
-#endif
-
-#if defined (HAVE_MPI)
-    printf("  Note: test can be also launched as an MPI application\n");
-    printf("\n");
-#elif defined (HAVE_RTE)
-    printf("  Note: this test can be also launched as an libRTE application\n");
-    printf("\n");
-#endif
-    printf("  Usage: %s [ server-hostname ] [ options ]\n", program);
-    printf("\n");
-    printf("  Common options:\n");
-    printf("     -t <test>      test to run:\n");
-    for (test = tests; test->name; ++test) {
-        printf("    %13s - %s %s\n", test->name,
-               api_names[test->api], test->desc);
-    }
-    printf("\n");
-    printf("     -s <size>      list of scatter-gather sizes for single message (%zu)\n",
-                                ctx->params.super.msg_size_list[0]);
-    printf("                    for example: \"-s 16,48,8192,8192,14\"\n");
-    printf("     -m <send mem type>[,<recv mem type>]\n");
-    printf("                    memory type of message for sender and receiver (host)\n");
-    print_memory_type_usage();
-    printf("     -n <iters>     number of iterations to run (%"PRIu64")\n", ctx->params.super.max_iter);
-    printf("     -w <iters>     number of warm-up iterations (%"PRIu64")\n",
-                                ctx->params.super.warmup_iter);
-    printf("     -c <cpulist>   set affinity to this CPU list (separated by comma) (off)\n");
-    printf("     -O <count>     maximal number of uncompleted outstanding sends\n");
-    printf("     -i <offset>    distance between consecutive scatter-gather entries (%zu)\n",
-                                ctx->params.super.iov_stride);
-    printf("     -T <threads>   number of threads in the test (%d)\n",
-                                ctx->params.super.thread_count);
-    printf("     -o             do not progress the responder in one-sided tests\n");
-    printf("     -B             register memory with NONBLOCK flag\n");
-    printf("     -b <file>      read and execute tests from a batch file: every line in the\n");
-    printf("                    file is a test to run, first word is test name, the rest of\n");
-    printf("                    the line is command-line arguments for the test.\n");
-    printf("     -R <rank>      percentile rank of the percentile data in latency tests (%.1f)\n",
-                                ctx->params.super.percentile_rank);
-    printf("     -p <port>      TCP port to use for data exchange (%d)\n", ctx->port);
-#ifdef HAVE_MPI
-    printf("     -P <0|1>       disable/enable MPI mode (%d)\n", ctx->mpi);
-#endif
-    printf("     -h             show this help message\n");
-    printf("\n");
-    printf("  Output format:\n");
-    printf("     -N             use numeric formatting (thousands separator)\n");
-    printf("     -f             print only final numbers\n");
-    printf("     -v             print CSV-formatted output\n");
-    printf("\n");
-    printf("  UCT only:\n");
-    printf("     -d <device>    device to use for testing\n");
-    printf("     -x <tl>        transport to use for testing\n");
-    printf("     -D <layout>    data layout for sender side:\n");
-    printf("                        short    - short messages (default, cannot be used for get)\n");
-    printf("                        shortiov - short io-vector messages (only for active messages)\n");
-    printf("                        bcopy    - copy-out (cannot be used for atomics)\n");
-    printf("                        zcopy    - zero-copy (cannot be used for atomics)\n");
-    printf("     -W <count>     flow control window size, for active messages (%u)\n",
-                                ctx->params.super.uct.fc_window);
-    printf("     -H <size>      active message header size (%zu), included in message size\n",
-                                ctx->params.super.uct.am_hdr_size);
-    printf("     -A <mode>      asynchronous progress mode (thread_spinlock)\n");
-    printf("                        thread_spinlock - separate progress thread with spin locking\n");
-    printf("                        thread_mutex - separate progress thread with mutex locking\n");
-    printf("                        signal - signal-based timer\n");
-    printf("\n");
-    printf("  UCP only:\n");
-    printf("     -M <thread>    thread support level for progress engine (single)\n");
-    printf("                        single     - only the master thread can access\n");
-    printf("                        serialized - one thread can access at a time\n");
-    printf("                        multi      - multiple threads can access\n");
-    printf("     -D <layout>[,<layout>]\n");
-    printf("                    data layout for sender and receiver side (contig)\n");
-    printf("                        contig - Continuous datatype\n");
-    printf("                        iov    - Scatter-gather list\n");
-    printf("     -C             use wild-card tag for tag tests\n");
-    printf("     -U             force unexpected flow by using tag probe\n");
-    printf("     -r <mode>      receive mode for stream tests (recv)\n");
-    printf("                        recv       : Use ucp_stream_recv_nb\n");
-    printf("                        recv_data  : Use ucp_stream_recv_data_nb\n");
-    printf("     -I             create context with wakeup feature enabled\n");
-    printf("     -e             create endpoints with error handling support\n");
-    printf("     -E <mode>      wait mode for tests\n");
-    printf("                        poll       : repeatedly call worker_progress\n");
-    printf("                        sleep      : go to sleep after posting requests\n");
-    printf("     -H <size>      active message header size (%zu), not included in message size\n",
-                                ctx->params.super.ucp.am_hdr_size);
-    printf("\n");
-    printf("   NOTE: When running UCP tests, transport and device should be specified by\n");
-    printf("         environment variables: UCX_TLS and UCX_[SELF|SHM|NET]_DEVICES.\n");
-    printf("\n");
-}
-
-static ucs_status_t parse_ucp_datatype_params(const char *opt_arg,
-                                              ucp_perf_datatype_t *datatype)
-{
-    const char  *iov_type         = "iov";
-    const size_t iov_type_size    = strlen("iov");
-    const char  *contig_type      = "contig";
-    const size_t contig_type_size = strlen("contig");
-
-    if (0 == strncmp(opt_arg, iov_type, iov_type_size)) {
-        *datatype = UCP_PERF_DATATYPE_IOV;
-    } else if (0 == strncmp(opt_arg, contig_type, contig_type_size)) {
-        *datatype = UCP_PERF_DATATYPE_CONTIG;
-    } else {
-        return UCS_ERR_INVALID_PARAM;
-    }
-
-    return UCS_OK;
-}
-
-static ucs_status_t parse_mem_type(const char *opt_arg,
-                                   ucs_memory_type_t *mem_type)
-{
-    ucs_memory_type_t it;
-    for (it = UCS_MEMORY_TYPE_HOST; it < UCS_MEMORY_TYPE_LAST; it++) {
-        if(!strcmp(opt_arg, ucs_memory_type_names[it]) &&
-           (ucx_perf_mem_type_allocators[it] != NULL)) {
-            *mem_type = it;
-            return UCS_OK;
-        }
-    }
-    ucs_error("Unsupported memory type: \"%s\"", opt_arg);
-    return UCS_ERR_INVALID_PARAM;
-}
-
-static ucs_status_t parse_mem_type_params(const char *opt_arg,
-                                          ucs_memory_type_t *send_mem_type,
-                                          ucs_memory_type_t *recv_mem_type)
-{
-    const char *delim = ",";
-    char *token       = strtok((char*)opt_arg, delim);
-
-    if (UCS_OK != parse_mem_type(token, send_mem_type)) {
-        return UCS_ERR_INVALID_PARAM;
-    }
-
-    token = strtok(NULL, delim);
-    if (NULL == token) {
-        *recv_mem_type = *send_mem_type;
-        return UCS_OK;
-    } else {
-        return parse_mem_type(token, recv_mem_type);
-    }
-}
-
-static ucs_status_t parse_message_sizes_params(const char *opt_arg,
-                                               ucx_perf_params_t *params)
-{
-    const char delim = ',';
-    size_t *msg_size_list, token_num, token_it;
-    char *optarg_ptr, *optarg_ptr2;
-
-    optarg_ptr = (char *)opt_arg;
-    token_num  = 0;
-    /* count the number of given message sizes */
-    while ((optarg_ptr = strchr(optarg_ptr, delim)) != NULL) {
-        ++optarg_ptr;
-        ++token_num;
-    }
-    ++token_num;
-
-    msg_size_list = realloc(params->msg_size_list,
-                            sizeof(*params->msg_size_list) * token_num);
-    if (NULL == msg_size_list) {
-        return UCS_ERR_NO_MEMORY;
-    }
-
-    params->msg_size_list = msg_size_list;
-
-    optarg_ptr = (char *)opt_arg;
-    errno = 0;
-    for (token_it = 0; token_it < token_num; ++token_it) {
-        params->msg_size_list[token_it] = strtoul(optarg_ptr, &optarg_ptr2, 10);
-        if (((ERANGE == errno) && (ULONG_MAX == params->msg_size_list[token_it])) ||
-            ((errno != 0) && (params->msg_size_list[token_it] == 0)) ||
-            (optarg_ptr == optarg_ptr2)) {
-            free(params->msg_size_list);
-            params->msg_size_list = NULL; /* prevent double free */
-            ucs_error("Invalid option substring argument at position %lu", token_it);
-            return UCS_ERR_INVALID_PARAM;
-        }
-        optarg_ptr = optarg_ptr2 + 1;
-    }
-
-    params->msg_size_cnt = token_num;
-    return UCS_OK;
-}
-
-static ucs_status_t init_test_params(perftest_params_t *params)
+ucs_status_t init_test_params(perftest_params_t *params)
 {
     memset(params, 0, sizeof(*params));
     params->super.api               = UCX_PERF_API_LAST;
@@ -640,351 +179,6 @@ static ucs_status_t init_test_params(perftest_params_t *params)
 
     params->super.msg_size_list[0] = 8;
     params->test_id                = TEST_ID_UNDEFINED;
-
-    return UCS_OK;
-}
-
-static ucs_status_t parse_test_params(perftest_params_t *params, char opt,
-                                      const char *opt_arg)
-{
-    char *optarg2 = NULL;
-    test_type_t *test;
-    unsigned i;
-
-    switch (opt) {
-    case 'd':
-        ucs_snprintf_zero(params->super.uct.dev_name,
-                          sizeof(params->super.uct.dev_name), "%s", opt_arg);
-        return UCS_OK;
-    case 'x':
-        ucs_snprintf_zero(params->super.uct.tl_name,
-                          sizeof(params->super.uct.tl_name), "%s", opt_arg);
-        return UCS_OK;
-    case 't':
-        for (i = 0; tests[i].name != NULL; ++i) {
-            test = &tests[i];
-            if (!strcmp(opt_arg, test->name)) {
-                params->super.api       = test->api;
-                params->super.command   = test->command;
-                params->super.test_type = test->test_type;
-                params->test_id         = i;
-                break;
-            }
-        }
-        if (params->test_id == TEST_ID_UNDEFINED) {
-            ucs_error("Invalid option argument for -t");
-            return UCS_ERR_INVALID_PARAM;
-        }
-        return UCS_OK;
-    case 'D':
-        if (!strcmp(opt_arg, "short")) {
-            params->super.uct.data_layout   = UCT_PERF_DATA_LAYOUT_SHORT;
-        } else if (!strcmp(opt_arg, "shortiov")) {
-            params->super.uct.data_layout   = UCT_PERF_DATA_LAYOUT_SHORT_IOV;
-        } else if (!strcmp(opt_arg, "bcopy")) {
-            params->super.uct.data_layout   = UCT_PERF_DATA_LAYOUT_BCOPY;
-        } else if (!strcmp(opt_arg, "zcopy")) {
-            params->super.uct.data_layout   = UCT_PERF_DATA_LAYOUT_ZCOPY;
-        } else if (UCS_OK == parse_ucp_datatype_params(opt_arg,
-                                                       &params->super.ucp.send_datatype)) {
-            optarg2 = strchr(opt_arg, ',');
-            if (optarg2) {
-                if (UCS_OK != parse_ucp_datatype_params(optarg2 + 1,
-                                                       &params->super.ucp.recv_datatype)) {
-                    return UCS_ERR_INVALID_PARAM;
-                }
-            }
-        } else {
-            ucs_error("Invalid option argument for -D");
-            return UCS_ERR_INVALID_PARAM;
-        }
-        return UCS_OK;
-    case 'E':
-        if (!strcmp(opt_arg, "poll")) {
-            params->super.wait_mode = UCX_PERF_WAIT_MODE_POLL;
-            return UCS_OK;
-        } else if (!strcmp(opt_arg, "sleep")) {
-            params->super.wait_mode = UCX_PERF_WAIT_MODE_SLEEP;
-            return UCS_OK;
-        } else {
-            ucs_error("Invalid option argument for -E");
-            return UCS_ERR_INVALID_PARAM;
-        }
-        return UCS_OK;
-    case 'i':
-        params->super.iov_stride = atol(opt_arg);
-        return UCS_OK;
-    case 'n':
-        params->super.max_iter = atol(opt_arg);
-        return UCS_OK;
-    case 's':
-        return parse_message_sizes_params(opt_arg, &params->super);
-    case 'H':
-        params->super.uct.am_hdr_size = atol(opt_arg);
-        params->super.ucp.am_hdr_size = atol(opt_arg);
-        return UCS_OK;
-    case 'W':
-        params->super.uct.fc_window = atoi(opt_arg);
-        return UCS_OK;
-    case 'O':
-        params->super.max_outstanding = atoi(opt_arg);
-        return UCS_OK;
-    case 'w':
-        params->super.warmup_iter = atol(opt_arg);
-        return UCS_OK;
-    case 'o':
-        params->super.flags |= UCX_PERF_TEST_FLAG_ONE_SIDED;
-        return UCS_OK;
-    case 'B':
-        params->super.flags |= UCX_PERF_TEST_FLAG_MAP_NONBLOCK;
-        return UCS_OK;
-    case 'q':
-        params->super.flags &= ~UCX_PERF_TEST_FLAG_VERBOSE;
-        return UCS_OK;
-    case 'C':
-        params->super.flags |= UCX_PERF_TEST_FLAG_TAG_WILDCARD;
-        return UCS_OK;
-    case 'U':
-        params->super.flags |= UCX_PERF_TEST_FLAG_TAG_UNEXP_PROBE;
-        return UCS_OK;
-    case 'I':
-        params->super.flags |= UCX_PERF_TEST_FLAG_WAKEUP;
-        return UCS_OK;
-    case 'e':
-        params->super.flags |= UCX_PERF_TEST_FLAG_ERR_HANDLING;
-        return UCS_OK;
-    case 'M':
-        if (!strcmp(opt_arg, "single")) {
-            params->super.thread_mode = UCS_THREAD_MODE_SINGLE;
-            return UCS_OK;
-        } else if (!strcmp(opt_arg, "serialized")) {
-            params->super.thread_mode = UCS_THREAD_MODE_SERIALIZED;
-            return UCS_OK;
-        } else if (!strcmp(opt_arg, "multi")) {
-            params->super.thread_mode = UCS_THREAD_MODE_MULTI;
-            return UCS_OK;
-        } else {
-            ucs_error("Invalid option argument for -M");
-            return UCS_ERR_INVALID_PARAM;
-        }
-    case 'T':
-        params->super.thread_count = atoi(opt_arg);
-        return UCS_OK;
-    case 'A':
-        if (!strcmp(opt_arg, "thread") || !strcmp(opt_arg, "thread_spinlock")) {
-            params->super.async_mode = UCS_ASYNC_MODE_THREAD_SPINLOCK;
-            return UCS_OK;
-        } else if (!strcmp(opt_arg, "thread_mutex")) {
-            params->super.async_mode = UCS_ASYNC_MODE_THREAD_MUTEX;
-            return UCS_OK;
-        } else if (!strcmp(opt_arg, "signal")) {
-            params->super.async_mode = UCS_ASYNC_MODE_SIGNAL;
-            return UCS_OK;
-        } else {
-            ucs_error("Invalid option argument for -A");
-            return UCS_ERR_INVALID_PARAM;
-        }
-    case 'r':
-        if (!strcmp(opt_arg, "recv_data")) {
-            params->super.flags |= UCX_PERF_TEST_FLAG_STREAM_RECV_DATA;
-            return UCS_OK;
-        } else if (!strcmp(opt_arg, "recv")) {
-            params->super.flags &= ~UCX_PERF_TEST_FLAG_STREAM_RECV_DATA;
-            return UCS_OK;
-        }
-        return UCS_ERR_INVALID_PARAM;
-    case 'R':
-        params->super.percentile_rank = atof(opt_arg);
-        if ((0.0 <= params->super.percentile_rank) && (params->super.percentile_rank <= 100.0)) {
-            return UCS_OK;
-        } else {
-            ucs_error("Invalid option argument for -R");
-            return UCS_ERR_INVALID_PARAM;
-        }
-    case 'm':
-        if (UCS_OK != parse_mem_type_params(opt_arg,
-                                            &params->super.send_mem_type,
-                                            &params->super.recv_mem_type)) {
-            return UCS_ERR_INVALID_PARAM;
-        }
-        return UCS_OK;
-    default:
-       return UCS_ERR_INVALID_PARAM;
-    }
-}
-
-static ucs_status_t adjust_test_params(perftest_params_t *params,
-                                       const char *error_prefix)
-{
-    test_type_t *test;
-
-    if (params->test_id == TEST_ID_UNDEFINED) {
-        ucs_error("%smissing test name", error_prefix);
-        return UCS_ERR_INVALID_PARAM;
-    }
-
-    test = &tests[params->test_id];
-
-    if (params->super.max_outstanding == 0) {
-        params->super.max_outstanding = test->window_size;
-    }
-
-    return UCS_OK;
-}
-
-static ucs_status_t read_batch_file(FILE *batch_file, const char *file_name,
-                                    int *line_num, perftest_params_t *params,
-                                    char** test_name_p)
-{
-#define MAX_SIZE 256
-#define MAX_ARG_SIZE 2048
-    ucs_status_t status;
-    char buf[MAX_ARG_SIZE];
-    char error_prefix[MAX_ARG_SIZE];
-    int argc;
-    char *argv[MAX_SIZE + 1];
-    int c;
-    char *p;
-
-    do {
-        if (fgets(buf, sizeof(buf) - 1, batch_file) == NULL) {
-            return UCS_ERR_NO_ELEM;
-        }
-        ++(*line_num);
-
-        argc = 0;
-        p = strtok(buf, " \t\n\r");
-        while (p && (argc < MAX_SIZE)) {
-            argv[argc++] = p;
-            p = strtok(NULL, " \t\n\r");
-        }
-        argv[argc] = NULL;
-    } while ((argc == 0) || (argv[0][0] == '#'));
-
-    ucs_snprintf_safe(error_prefix, sizeof(error_prefix),
-                      "in batch file '%s' line %d: ", file_name, *line_num);
-
-    optind = 1;
-    while ((c = getopt (argc, argv, TEST_PARAMS_ARGS)) != -1) {
-        status = parse_test_params(params, c, optarg);
-        if (status != UCS_OK) {
-            ucs_error("%s-%c %s: %s", error_prefix, c, optarg,
-                      ucs_status_string(status));
-            return status;
-        }
-    }
-
-    status = adjust_test_params(params, error_prefix);
-    if (status != UCS_OK) {
-        return status;
-    }
-
-    *test_name_p = strdup(argv[0]);
-    return UCS_OK;
-}
-
-static ucs_status_t parse_cpus(char *opt_arg, struct perftest_context *ctx)
-{
-    char *endptr, *cpu_list = opt_arg;
-    int cpu;
-
-    ctx->num_cpus = 0;
-    cpu           = strtol(cpu_list, &endptr, 10);
-
-    while (((*endptr == ',') || (*endptr == '\0')) && (ctx->num_cpus < MAX_CPUS)) {
-        if (cpu < 0) {
-            ucs_error("invalid cpu number detected: (%d)", cpu);
-            return UCS_ERR_INVALID_PARAM;
-        }
-
-        ctx->cpus[ctx->num_cpus++] = cpu;
-
-        if (*endptr == '\0') {
-            break;
-        }
-
-        cpu_list = endptr + 1; /* skip the comma */
-        cpu      = strtol(cpu_list, &endptr, 10);
-    }
-
-    if (*endptr == ',') {
-        ucs_error("number of listed cpus exceeds the maximum supported value (%d)",
-                  MAX_CPUS);
-        return UCS_ERR_INVALID_PARAM;
-    }
-
-    return UCS_OK;
-}
-
-static ucs_status_t parse_opts(struct perftest_context *ctx, int mpi_initialized,
-                               int argc, char **argv)
-{
-    ucs_status_t status;
-    int c;
-
-    ucs_trace_func("");
-
-    ucx_perf_global_init(); /* initialize memory types */
-
-    status = init_test_params(&ctx->params);
-    if (status != UCS_OK) {
-        return status;
-    }
-
-    ctx->server_addr            = NULL;
-    ctx->num_batch_files        = 0;
-    ctx->port                   = 13337;
-    ctx->flags                  = 0;
-    ctx->mpi                    = mpi_initialized;
-
-    optind = 1;
-    while ((c = getopt (argc, argv, "p:b:Nfvc:P:h" TEST_PARAMS_ARGS)) != -1) {
-        switch (c) {
-        case 'p':
-            ctx->port = atoi(optarg);
-            break;
-        case 'b':
-            if (ctx->num_batch_files < MAX_BATCH_FILES) {
-                ctx->batch_files[ctx->num_batch_files++] = optarg;
-            }
-            break;
-        case 'N':
-            ctx->flags |= TEST_FLAG_NUMERIC_FMT;
-            break;
-        case 'f':
-            ctx->flags |= TEST_FLAG_PRINT_FINAL;
-            break;
-        case 'v':
-            ctx->flags |= TEST_FLAG_PRINT_CSV;
-            break;
-        case 'c':
-            ctx->flags |= TEST_FLAG_SET_AFFINITY;
-            status = parse_cpus(optarg, ctx);
-            if (status != UCS_OK) {
-                return status;
-            }
-            break;
-        case 'P':
-#ifdef HAVE_MPI
-            ctx->mpi = atoi(optarg) && mpi_initialized;
-            break;
-#endif
-        case 'h':
-            usage(ctx, ucs_basename(argv[0]));
-            return UCS_ERR_CANCELED;
-        default:
-            status = parse_test_params(&ctx->params, c, optarg);
-            if (status != UCS_OK) {
-                usage(ctx, ucs_basename(argv[0]));
-                return status;
-            }
-            break;
-        }
-    }
-
-    if (optind < argc) {
-        ctx->server_addr = argv[optind];
-    }
 
     return UCS_OK;
 }
@@ -1594,142 +788,6 @@ static ucs_status_t check_system(struct perftest_context *ctx)
     }
 
     return UCS_OK;
-}
-
-static ucs_status_t clone_params(perftest_params_t *dest,
-                                 const perftest_params_t *src)
-{
-    size_t msg_size_list_size;
-
-    *dest                     = *src;
-    msg_size_list_size        = dest->super.msg_size_cnt *
-                                sizeof(*dest->super.msg_size_list);
-    dest->super.msg_size_list = malloc(msg_size_list_size);
-    if (dest->super.msg_size_list == NULL) {
-        return ((msg_size_list_size != 0) ? UCS_ERR_NO_MEMORY : UCS_OK);
-    }
-
-    memcpy(dest->super.msg_size_list, src->super.msg_size_list,
-           msg_size_list_size);
-    return UCS_OK;
-}
-
-static ucs_status_t check_params(const perftest_params_t *params)
-{
-    switch (params->super.api) {
-    case UCX_PERF_API_UCT:
-        if (!strcmp(params->super.uct.dev_name, TL_RESOURCE_NAME_NONE)) {
-            ucs_error("A device must be specified with -d flag for UCT test");
-            return UCS_ERR_INVALID_PARAM;
-        }
-        if (!strcmp(params->super.uct.tl_name, TL_RESOURCE_NAME_NONE)) {
-            ucs_error(
-                    "A transport must be specified with -x flag for UCT test");
-            return UCS_ERR_INVALID_PARAM;
-        }
-        return UCS_OK;
-    case UCX_PERF_API_UCP:
-        if (strcmp(params->super.uct.dev_name, TL_RESOURCE_NAME_NONE)) {
-            ucs_warn("-d '%s' ignored for UCP test; see NOTES section in help "
-                     "message",
-                     params->super.uct.dev_name);
-        }
-        if (strcmp(params->super.uct.tl_name, TL_RESOURCE_NAME_NONE)) {
-            ucs_warn("-x '%s' ignored for UCP test; see NOTES section in help "
-                     "message",
-                     params->super.uct.tl_name);
-        }
-        return UCS_OK;
-    default:
-        ucs_error("Invalid test case");
-        return UCS_ERR_INVALID_PARAM;
-    }
-}
-
-static ucs_status_t run_test_recurs(struct perftest_context *ctx,
-                                    const perftest_params_t *parent_params,
-                                    unsigned depth)
-{
-    perftest_params_t params;
-    ucx_perf_result_t result;
-    ucs_status_t status;
-    FILE *batch_file;
-    int line_num;
-
-    ucs_trace_func("depth=%u, num_files=%u", depth, ctx->num_batch_files);
-
-    if (depth >= ctx->num_batch_files) {
-        print_test_name(ctx);
-        status = check_params(parent_params);
-        if (status != UCS_OK) {
-            return status;
-        }
-
-        return ucx_perf_run(&parent_params->super, &result);
-    }
-
-    batch_file = fopen(ctx->batch_files[depth], "r");
-    if (batch_file == NULL) {
-        ucs_error("Failed to open batch file '%s': %m", ctx->batch_files[depth]);
-        return UCS_ERR_IO_ERROR;
-    }
-
-    line_num = 0;
-    do {
-        status = clone_params(&params, parent_params);
-        if (status != UCS_OK) {
-            goto out;
-        }
-
-        status = read_batch_file(batch_file, ctx->batch_files[depth],
-                                 &line_num, &params,
-                                 &ctx->test_names[depth]);
-        if (status == UCS_OK) {
-            run_test_recurs(ctx, &params, depth + 1);
-            free(ctx->test_names[depth]);
-            ctx->test_names[depth] = NULL;
-        }
-
-        free(params.super.msg_size_list);
-        params.super.msg_size_list = NULL;
-    } while (status == UCS_OK);
-
-    if (status == UCS_ERR_NO_ELEM) {
-        status = UCS_OK;
-    }
-
-out:
-    fclose(batch_file);
-    return status;
-}
-
-static ucs_status_t run_test(struct perftest_context *ctx)
-{
-    const char *error_prefix;
-    ucs_status_t status;
-
-    ucs_trace_func("");
-
-    setlocale(LC_ALL, "en_US");
-
-    /* no batch files, only command line params */
-    if (ctx->num_batch_files == 0) {
-        error_prefix = (ctx->flags & TEST_FLAG_PRINT_RESULTS) ?
-                       "command line: " : "";
-        status       = adjust_test_params(&ctx->params, error_prefix);
-        if (status != UCS_OK) {
-            return status;
-        }
-    }
-
-    print_header(ctx);
-
-    status = run_test_recurs(ctx, &ctx->params, 0);
-    if (status != UCS_OK) {
-        ucs_error("Failed to run test: %s", ucs_status_string(status));
-    }
-
-    return status;
 }
 
 int main(int argc, char **argv)
